Several changes:
 - UpdatePasswordWidget does not delete itself when done but emit a signal. (It is useful when the widget is not in a dialog)
 - int role was still used in some places (instead of ItemDataRole)
 - refactorings and changes for JWt 4
 - Fixing access modifiers that Java is stricter about

+++ src/Wt/WTimer.C
@@ -56,7 +56,7 @@ void WTimer::start()
   }
 
   active_ = true;
-  *timeout_ = Time() + interval_.count();
+  *timeout_ = Time() + static_cast<int>(interval_.count());
 
   bool jsRepeat = !singleShot_ &&
                   ((app && app->environment().ajax()) ||
@@ -86,7 +86,7 @@ void WTimer::gotTimeout()
 {
   if (active_) {
     if (!singleShot_) {
-      *timeout_ = Time() + interval_.count();
+      *timeout_ = Time() + static_cast<int>(interval_.count());
       if (!timerWidget_->jsRepeat()) {
         WApplication *app = WApplication::instance();
         timerWidget_->timerStart(app->environment().ajax());
